refactor: application 層内に state と usecase を配置
